requirements.txt
streamlit==1.31.1
requests==2.31.0
huggingface-hub==0.31.1
python-dotenv==1.0.0
